"i2c registers factored out of HAL; updates #1173"
