+++ software/controller/lib/hal/i2c.h
@@ -22,10 +22,110 @@ limitations under the License.
 #include "circular_buffer.h"
 #include "dma.h"
 
-#if defined(BARE_STM32)
-/// \TODO: for i2c regs
-#include "hal_stm32_regs.h"
-#endif
+// [RM] 37.7 I2C Registers
+struct I2CStruct {
+  union {
+    struct {
+      uint32_t enable : 1;
+      uint32_t tx_interrupts : 1;
+      uint32_t rx_interrupts : 1;
+      uint32_t addr_interrupts : 1;
+      uint32_t nack_interrupts : 1;
+      uint32_t stop_interrupts : 1;
+      uint32_t tx_complete_interrupts : 1;
+      uint32_t error_interrupts : 1;
+      uint32_t digital_noise_filter_conntrol : 4;
+      uint32_t analog_noise_filternig_off : 1;
+      uint32_t reserved1 : 1;
+      uint32_t dma_tx : 1;
+      uint32_t dma_rx : 1;
+      uint32_t slave_byte : 1;
+      uint32_t no_clock_stretch : 1;
+      uint32_t wake_from_stop : 1;
+      uint32_t general_call : 1;
+      uint32_t smbus_default_host : 1;
+      uint32_t smbus_default_device : 1;
+      uint32_t smbus_alert : 1;
+      uint32_t smbus_packet_error_checking : 1;
+      uint32_t reserved2 : 8;
+    };
+    uint32_t full_reg;
+  } control_reg1;  // Control register 1 [RM] 37.7.1
+  union {
+    struct {
+      uint32_t slave_addr_lsb : 1;  // lsb of 10 bits slave address
+      // (master)
+      uint32_t slave_addr_7b : 7;  // middle 7b bits of slave address
+      // (master)
+      uint32_t slave_addr_msb : 2;  // msb of 10 bits slave address
+      // (master)
+      uint32_t transfer_direction : 1;  // 0 = write, 1 = read (master)
+      uint32_t address_10b : 1;         // Set to enable 10 bits address header
+      // (master)
+      uint32_t read_10b_header : 1;  // Clear to send complete read sequence
+      // (master)
+      uint32_t start : 1;  // Set to generate START condition
+      uint32_t stop : 1;   // Set to generate STOP after byte transfer
+      // (master)
+      uint32_t nack : 1;                     // Generate NACK after byte reception (slave)
+      uint32_t n_bytes : 8;                  // Set to the number of bytes to send
+      uint32_t reload : 1;                   // Set to allow several consecutive transfers
+      uint32_t autoend : 1;                  // Set to automatically send stop condition
+      uint32_t packet_error_check_byte : 1;  // Set to send SMBus packet error
+      // checking byte
+      uint32_t reserved : 5;
+    };
+    uint32_t full_reg;
+  } control2;        // Control register 2 [RM] 37.7.2
+  uint32_t addr[2];  // Own address (1-2) register [RM] 37.7.{3,4} (slave)
+  union {
+    struct {
+      uint32_t scl_low : 8;   // Duration of SCL Low state (cycles)
+      uint32_t scl_high : 8;  // Duration of SCL High state (cycles)
+      uint32_t sda_hold : 4;  // Delay between SCL falling edge and SDA
+      // edge (cycles)
+      uint32_t scl_delay : 4;  // Delay between SDA edge and SCL rising
+      // edge (cycles)
+      uint32_t reserved : 4;
+      uint32_t prescaler : 4;  // Prescaler
+    };
+    uint32_t full_reg;
+  } timing;          // Timing register [RM] 37.7.5
+  uint32_t timeout;  // Timout register [RM] 37.7.6
+  union {
+    struct {
+      uint32_t tx_empty : 1;
+      uint32_t tx_interrupt : 1;
+      uint32_t rx_not_empty : 1;
+      uint32_t address_match : 1;
+      uint32_t nack : 1;
+      uint32_t stop : 1;
+      uint32_t transfer_complete : 1;
+      uint32_t transfer_reload : 1;
+      uint32_t bus_error : 1;
+      uint32_t arbitration_loss : 1;
+      uint32_t overrun : 1;
+      uint32_t packet_check_error : 1;
+      uint32_t timeout : 1;
+      uint32_t alert : 1;
+      uint32_t reserved1 : 1;
+      uint32_t busy : 1;
+      uint32_t transfer_direction : 1;
+      uint32_t address_code : 7;
+      uint32_t reserved : 8;
+    };
+    uint32_t full_reg;
+  } status;                     // Interrupt & status register [RM] 37.7.7
+  uint32_t interrupt_clear;     // Interrupt clear register [RM] 37.7.8
+  uint32_t packet_error_check;  // PEC register [RM] 37.7.9
+  uint32_t rx_data;             // Receive data register [RM] 37.7.10
+  uint32_t tx_data;             // Transmit data register [RM] 37.7.11
+};
+typedef volatile I2CStruct I2CReg;
+inline I2CReg *const I2C1Base = reinterpret_cast<I2CReg *>(0x40005400);
+inline I2CReg *const I2C2Base = reinterpret_cast<I2CReg *>(0x40005800);
+inline I2CReg *const I2C3Base = reinterpret_cast<I2CReg *>(0x40005c00);
+inline I2CReg *const I2C4Base = reinterpret_cast<I2CReg *>(0x40008400);
 
 namespace I2C {
